refactor: Update Swiper.vue and Type.vue styling
This commit updates the styling of the Swiper.vue and Type.vue components. In Swiper.vue, a container div with the class "container-fluid" is added to wrap the swiper-container element. This improves the layout and responsiveness of the swiper component. In Type.vue, the slides-per-view property is changed from 3 to 2, and the height of the swiper-bg-2 class is reduced from 500px to 200px. These changes enhance the visual appearance and user experience of the application.