refactor: Update Type.vue to dynamically render types in Swiper component
This commit modifies the Type.vue component to dynamically render the types in the Swiper component. The axios GET request in the getApi() method now retrieves the types from the API and assigns them to the store.types property. The Swiper component is then updated to use v-for to iterate over the store.types array and render the types dynamically. This improves the flexibility and scalability of the application by allowing the types to be easily updated and managed through the API.