refactor: Improve styling and lazy load images in Type.vue component
This commit refactors the Type.vue component to enhance its styling and implement lazy loading for images. The font size of the navigation links in the Header.vue component is increased to 1.7rem for better visibility. The color of the navigation links is changed to #f6b802 when hovered or active. In the Hero.vue component, the width of the image is reduced to 30% and a negative margin-top of 50px is added for better alignment. The Swiper.vue component is updated to use a new image "swiper-test.png" for all slides. Additionally, the speed of the swiper animation is changed to 5000ms. The Type.vue component is modified to use checkboxes for types in the Swiper component. The "desk-img" and "tab-img" classes are added to the respective image elements to control their display on different screen sizes. The margin-bottom property of the "desk-img" class is adjusted for better alignment. These changes improve the performance and visual appeal of the Type component, providing a better user experience.