refactor: Update Card.vue, Type.vue, and Swiper.vue components
This commit refactors the Card.vue, Type.vue, and Swiper.vue components to improve their functionality and styling. In the Card.vue component, the imports are updated, the code is reformatted, and the template and styles are modified. In the Type.vue component, the styling is improved and lazy loading for images is implemented. The font size of the navigation links in the Header.vue component is increased to 1.7rem for better visibility, and the color of the navigation links is changed to #f6b802 when hovered or active. In the Hero.vue component, the width of the image is reduced to 30% and a negative margin-top of 50px is added for better alignment. The Swiper.vue component is updated to use a new image "swiper-test.png" for all slides, and the speed of the swiper animation is changed to 5000ms. These changes enhance the performance and visual appeal of the components, providing a better user experience.