refactor: Update Swiper.vue to improve layout and add background color
This commit modifies the Swiper.vue component to improve the layout by adding a new container element with a background color. The new container element has the class "swiper-bg" and is wrapped around the existing container element with the class "swiper-box". This change enhances the visual appeal of the component and provides a better separation between the swiper content and the background.